perf: improving a code perfs
