+++ REGISTER_MS/src/adapters/handlers/index.ts
@@ -1,2 +1,2 @@
-export * from './HandlerCreateUser'
-export * from './HandlerDeleteAllUsers'
+export *  from './HandlerCreateUser'
+export *  from './HandlerDeleteAllUsers'
