+++ REGISTER_MS/src/adapters/repository/index.ts
@@ -1,1 +1,1 @@
-export * from './UserRepository'
+export * from './PrismaUserRepository'
