+++ REGISTER_MS/src/domain/exceptions/index.ts
