+++ REGISTER_MS/src/domain/repository/index.ts
@@ -1,0 +1,1 @@
+export * from './UserRepository'
