--- REGISTER_MS/src/handlers/index.ts
@@ -1,2 +1,0 @@
-export *  from './HandlerCreateUser'
-export *  from './HandlerDeleteAllUsers'
