+++ REGISTER_MS/src/schemas/deleteAllUsers.schema.ts
@@ -1,6 +1,6 @@
 import * as yup from 'yup'
 
-const secret = process.env.DELETE_SECRET ?? 'delete-sem-where'
+const secret = process.env.DELETE_SECRET
 
 const deleteAllUsersSchema = yup.object({
   rootSecret: yup.string().required().equals([secret], 'get out of here!')
